Contacts, error messages

+++ src/services/contactsList.service.js
@@ -1,4 +1,4 @@
-import { UserNotFound } from '../errors/error-exceptions.js'
+import { ElementNotFound, UserAlreadyExists, UserNotFound } from '../errors/error-exceptions.js'
 import { contactsListRepository, usersRepository } from '../repositories/index.js'
 
 
@@ -7,25 +7,30 @@ export default class ContactsListService {
 
     getContacts = async (userId) => {
         const user =  await usersRepository.getById(userId)
-        const userList = user.contact_list
+        //const userList = user.contact_list
 
         if(!user){
-            throw new UserNotFound(`Users not found`)
+            throw new UserNotFound(`Usuario no encontrado`)
         }
-        const users = await contactsListRepository.getContacts(userList)
+        const contacts = await contactsListRepository.getContacts(userId)
 
-        return users
+        if(contacts.length === 0){
+            throw new ElementNotFound(`No hay contactos agregados aún`)
+        }
+
+        return contacts
     }
 
     getRequests = async (userId) => {
-        const user =  await usersRepository.getById(userId)
-        const userList = user.contact_list
+        //const user =  await usersRepository.getById(userId)
+        //const listId = user.contact_list
 
-        const requesters = await contactsListRepository.getRequests(userList)
+        const requesters = await contactsListRepository.getRequests(userId)
         
-        if(!requesters){
-            throw new UserNotFound(`No request yet`)
+        if(requesters.length === 0){
+            throw new ElementNotFound(`El usuario aún no tiene solicitudes de contacto`)
         }
+        
         return requesters
     }
 
@@ -33,7 +38,7 @@ export default class ContactsListService {
         const user =  await usersRepository.getById(userId)
         
         if(!user){
-            throw new UserNotFound(`Owner with ID N°${user._id} not found`)
+            throw new UserNotFound(`Usuario con Id N°${user._id} no encontrado`)
         }
 
         const result = await contactsListRepository.createList(userId)
@@ -41,63 +46,60 @@ export default class ContactsListService {
         return result
     }
 
-    addContact = async (userId, contactId) => {
-        const user = await usersRepository.getById(userId)
-        const userList = user.contact_list
-        
-        if(!userList){
-            throw new UserNotFound(`User with ID N° ${userId} not found`)
-        }
-        const receptor = await usersRepository.getById(contactId)
-        const receptorList = receptor.contact_list
-        
-        if(!receptorList){
-            throw new UserNotFound(`User with ID N° ${contactId} not found`)
+    addContact = async (userId, contactId) => {        
+        const contact = await contactsListRepository.isContact(userId, contactId)
+
+        if(contact){
+            throw new UserAlreadyExists(`El usuario ya pertenece a la lista de contactos`)
         }
 
-        const result = await contactsListRepository.addContact(userList, contactId)
-        await contactsListRepository.sendRequest(receptorList, userId)
+        const contacExists = await usersRepository.getById(contactId)
+        
+        if(!contacExists){
+            throw new UserNotFound(`Usuario con Id N°${contactId} no encontrado`)
+        }
+
+        const result = await contactsListRepository.addContact(userId, contactId)
+        await contactsListRepository.addContact(contactId, userId)
 
         return result
-
     }
 
     acceptContact = async (userId, contactId) => {
         const user = await usersRepository.getById(userId)
-        const userList = user.contact_list
 
         if(!user){
-            throw new UserNotFound(`Owner with ID N° ${user} not found`)
+            throw new UserNotFound(`Usuario con Id N°${user} no encontrado`)
         }
 
         const receptor = await usersRepository.getById(contactId)
-        const receptorList = receptor.contact_list
 
         if(!receptor){
-            throw new UserNotFound(`Contact with ID N° ${contactId} not found`)
+            throw new UserNotFound(`Usuario con Id N°${contactId} no encontrado`)
         }
 
-        await contactsListRepository.acceptContact(userList, contactId)
-        await contactsListRepository.acceptContact(receptorList, userId)
+        await contactsListRepository.acceptContact(userId, contactId)
+        await contactsListRepository.acceptContact(contactId, userId)
     }
 
     deleteContact = async(userId, contactId) => {
         const user = await usersRepository.getById(userId)
-        const userList = user.contact_list
-
-        const contactData = await usersRepository.getById(contactId)
-        const contactList = contactData.contact_list
+        const contactIsUser = await usersRepository.getById(contactId)
 
         if(!user){
-            throw new UserNotFound(`Owner with ID N° ${owner} not found`)
+            throw new UserNotFound(`Usuario con Id N°${user} no encontrado`)
+        }
+
+        if(!contactIsUser){
+            throw new UserNotFound(`El contacto Id N°${user} no es un usuario`)
         }
         
-        const contact = await contactsListRepository.getContact(userList, contactId)
+        const contact = await contactsListRepository.isContact(userId, contactId)
 
         if(!contact){
-            throw new UserNotFound(`Contact with ID N° ${contactId} not found`)
+            throw new UserNotFound(`Usuario con Id N°${contactId} no encontrado`)
         }
-        await contactsListRepository.deleteContact(contactList, userId)
+        await contactsListRepository.deleteContact(userId, userId)
 
         const result = await contactsListRepository.deleteContact(userList, contactId)
